refactor mappers.js
