+++ packages/tc-ui/src/primitives/server.js
@@ -1,4 +1,3 @@
-const sortBy = require('lodash.sortby');
 const Text = require('./text/server');
 const Boolean = require('./boolean/server');
 const Enum = require('./enum/server');
@@ -15,49 +14,6 @@ const addDefaults = obj => ({
 	...obj,
 });
 
-const getValue = (itemSchema, itemValue) => {
-	// preserves boolean values to prevent false being coerced to empty string
-	if (itemSchema.type === 'Boolean') {
-		return typeof itemValue === 'boolean' ? itemValue : '';
-	}
-
-	// return relationships as type, code and name object
-	if (itemSchema.relationship) {
-		if (itemSchema.hasMany) {
-			return itemValue
-				? sortBy(itemValue, `${itemSchema.type}.code`).map(item =>
-						Object.assign(
-							{},
-							{
-								code: item.code || item[itemSchema.type].code,
-								name:
-									item.name ||
-									item.code ||
-									item[itemSchema.type].name ||
-									item[itemSchema.type].code,
-							},
-							...itemValue,
-						),
-				  )
-				: [];
-		}
-		return itemValue
-			? {
-					code: itemValue.code || itemValue[itemSchema.type].code,
-					name:
-						itemValue.name ||
-						itemValue.code ||
-						itemValue[itemSchema.type].name ||
-						itemValue[itemSchema.type].code,
-					...itemValue,
-			  }
-			: null;
-	}
-
-	// everything else is just text
-	return itemValue;
-};
-
 module.exports = {
 	Text: addDefaults(Text),
 	Boolean: addDefaults(Boolean),
@@ -67,5 +23,4 @@ module.exports = {
 	Temporal: addDefaults(Temporal),
 	Relationship: addDefaults(Relationship),
 	MultipleChoice: addDefaults(MultipleChoice),
-	getValue,
 };
