+++ packages/tc-ui/src/server.js
@@ -1,6 +1,7 @@
 const primitives = require('./primitives/server');
-
-const { componentAssigner, graphqlQueryBuilder } = require('./lib/mappers');
+const componentAssigner = require('./lib/mappers/component-assigner');
+const graphqlQueryBuilder = require('./lib/mappers/graphql-query-builder');
+// const { componentAssigner, graphqlQueryBuilder } = require('./lib/mappers.js');
 const { ApiClient } = require('./lib/api-client');
 const { getSchemaSubset } = require('./lib/get-schema-subset');
 const { getDataTransformers } = require('./lib/get-data-transformers');
